legal-ai-assistant/requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
python-docx==1.1.0
PyPDF2==3.0.1